Many enhancements.
This addresses a number of the initial concerns. It seperates out
each sensor to its own node.  Some data from the device status
is added (RSSI and uptime).  Battery voltage for each sensor is
now listed. Some additional derived data is include like
dewpoint and apparent temperature and pressure trend.

Added a timer to keep track of how long it's been since an
update was recieved from the sensor (in minutes).

Started work on the configuration web page. It will display
a page, but doesn't process any of the config info yet.