Configuration updates
Generate a correct configuration web page and process changes
made to the configuration.

Add a static class to hold all the configuration data. instead of
indiviaual variables.

Add ability to write the configuration data as a json file and
the ability to read the json file and set the configuration data.

Updating the configuration on the webpage saves the configuration
data.

Move the event classes into a separate file to make the main file
a bit smaller and easier to maintain.

Add support for a heartbeat node. This will be a simple node that
has one value that toggles between -1 and 1. The ISY can check
this with a program to determine if the node server is still
alive.