Big overhall
The main goal of these changes is to get the web based
configuration working properly. Most of the changes
are to the web page generation and processes with tweaks
to the configuration system and the various startup
processes to make them restartable.

As part of this, it is now possible to query the Weather
Flow servers to get station meta data (elevation mostly).