Add serial_number property.
This returns the serial number as it appears in the data stream from
the device.  Previously we only wanted the serial number modified
to meet the naming conventions for the ISY node address but soon
we'll want to use this as a key and compare it with the incoming
data.