Upload profiles to ISY automatically.
Embed the profile files in the application and track the version
so that any updated files can be uploaded to the ISY when the
node server starts.

Unfortunately, this still seems to require an ISY reboot for any
changed files to take effect.

In addition, add a function to query the ISY for the installed
node servers and use the returned list to find our profile number.
This should be a better method of getting the profile than what
was previously used.