Restart/Pause server proceses from web interface.
Being able to restart the node server implies that it should be
possible to stop (or pause) it.  But the restart button only
attempts to stop/restart the processes.

Change the code so that we have both a restart and a pause
button.  To make it obvious what the current state is, add a
text field that switches between "running" and "paused" showing
the current state.

Do some refactoring of the heartbeat code and the udp code
to make starting/stopping of these processes easier to handle.

Also consolidate all the functions that do the initial
communication with the ISY into a single call and do this when
the configuration changes.  Because this is just configuration
related, there's no reason to "pause/restart" anything here so
only re-run the initialization if the configuration changes.