Add locking to prevent race.
It may be possible to have an Update event try to resent the
counter while the timer is sending the current time since last
update. In this case, the resetting of the counter may be
overwritten by the timer update, causing the counter to increase
instead of reset.

Add some locking around these two areas so that they wait on
each other and serialize the updating of the counter.