Fix node definition uom.
The unit of measure (uom) attached to the data sent to the ISY
needs to match the uom assigned to a node value via the editor
associated with that value.

If they don't match, program conditions don't work.  I.E. you
can't compare an index with an speed.

Add a bunch of new editor types with the proper uom to match
what we sending and use them for the node definitions.

Also, this means we do need separate nodes for lightning
and rapid wind since the units are now different for metric
vs. imperial.