Change air lightning sensor status string.
There are a couple of lightning sensor status values that don't
really indicate a failure when set.  The lightning Noise and
lightning disturber status really just means that the sensor
is detecting signals that aren't lightning.  So change the
wording to "triggerd" instead of "failed".

Also fix up some of the node descriptions;
 s/Weather Flow/WeatherFlow/
 s/diagnostic/device/