Better log viewing.
Add a log server process. This creates a websocket server that
forwards log messages over the websocket.

Update the viewer web page to connect to the websocket to get
log messages.

Changes to try and limit how much of a load this puts on the web
browser by:

1) when connecting to the websocket server, the server will only
   send the latest 500 lines of log information. This can take
   a couple of seconds, trying to send a large log would take
   minutes.
2) only display a maximum of 8000 lines of log information. The
   most recent 8000 lines are displayed.  Processing this many
   lines will cause performance issues on the browser.

The limits on the browser display doesn't effect the actual log
size. If the log is saved, it will save the entire log. Clearing
the log, however, does actually clear out the internal log, not
just the displayed lines.